refactor: introduce some error-handling neatness
It’s neater to have a try/catch wrapper around async/await functions: in this case it can save a level of nesting for every controller file we have. For now though, there’s only this one controller using the wrapper, but more will come. The catch-all error controller can’t use this wrapper because error controllers in Express need a different function signature to work (4 arguments instead of 3, where the extra one is an error).