refactor: change how we test the catch-all controller
This way and the old way are both a bit smelly, but this change at least makes it explicit that we’re targeting that controller with a non-prod-only route, rather than trying to bend-over backwards to throw an error within the catch-all controller itself. Hopefully this new way makes more sense to someone else (or me later) reading it.